[Status Page] Preload data
